# openapi/taskmanager.yaml
openapi: 3.0.3
security: []
info:
  title: Task Manager API – Curso API First
  description: API de referencia para el curso (definición OpenAPI + implementación Express).
  contact:
    name: Equipo del curso
    email: soporte@taskmanager.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:3000
  description: Servidor local de desarrollo
# - url: https://staging.api.taskmanager.dev/v1
#   description: Servidor de staging
tags:
- name: Tasks
  description: Operaciones relacionadas con la gestión de tareas
paths:
  /tasks:
    get:
      operationId: getTasks
      tags:
      - Tasks
      summary: Obtener todas las tareas
      description: Devuelve una lista de tareas registradas
      parameters:
      - in: query
        name: page
        schema:
          type: integer
          default: 1
          minimum: 1
        description: Página de resultados
      - in: query
        name: limit
        schema:
          type: integer
          default: 10
          minimum: 1
          maximum: 100
        description: Número de resultados por página
      - in: query
        name: status
        schema:
          type: string
          enum: [todo, in_progress, done]
        description: Filtrar por estado
      - in: query
        name: priority
        schema:
          type: string
          enum: [low, medium, high]
        description: Filtrar por prioridad
      - in: query
        name: search
        schema:
          type: string
        description: Buscar texto en el título de la tarea
      # responses:
      #   "200":
      #     description: Lista de tareas
      #     content:
      #       application/json:
      #         schema:
      #           type: array
      #           items:
      #             $ref: "#/components/schemas/Task"
      # responses:
      #   '200':
      #     description: Lista de tareas
      #     content:
      #       application/json:
      #         examples:
      #           ejemplo:
      #             value:
      #             - id: "1"
      #               title: "Mock desde OpenAPI"
      #               status: "todo"
      #             - id: "2"
      #               title: "nueva tarea Mock"
      #               status: "in_progress"
      responses:
        '200':
          description: Lista paginada de tareas
          content:
            application/json:
              schema:
                type: object
                required: [items, page, limit, totalItems]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  page:
                    type: integer
                    example: 1
                  limit:
                    type: integer
                    example: 10
                  totalItems:
                    type: integer
                    example: 42
                  totalPages:
                    type: integer
                    example: 5
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: createTask
      tags:
      - Tasks
      summary: Crear una nueva tarea
      description: Crea una nueva tarea a partir de los datos enviados
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/tasks_body"
        required: true
      responses:
        "201":
          description: Tarea creada correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /tasks/{id}:
    get:
      operationId: getTaskById
      tags:
      - Tasks
      summary: Obtener una tarea por ID
      description: Devuelve una tarea específica por su identificador
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: ID de la tarea
      responses:
        '200':
          description: Tarea encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Tarea no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ErrorResponse'
    put:
      operationId: updateTask
      tags:
      - Tasks
      summary: Actualizar una tarea completa
      description: Reemplaza completamente una tarea existente
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: ID de la tarea
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/tasks_body"
        required: true
      responses:
        '200':
          description: Tarea actualizada correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '404':
          description: Tarea no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ErrorResponse'
    patch:
      operationId: patchTask
      tags:
      - Tasks
      summary: Actualizar una tarea parcialmente
      description: Actualiza solo los campos especificados de una tarea
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: ID de la tarea
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                status:
                  type: string
                  enum: [todo, in_progress, done]
                priority:
                  type: string
                  enum: [low, medium, high]
                description:
                  type: string
                dueDate:
                  type: string
                  format: date-time
        required: true
      responses:
        '200':
          description: Tarea actualizada correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '404':
          description: Tarea no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      operationId: deleteTask
      tags:
      - Tasks
      summary: Eliminar una tarea
      description: Elimina una tarea del sistema
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: ID de la tarea
      responses:
        '204':
          description: Tarea eliminada correctamente
        '404':
          description: Tarea no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ErrorResponse'
components:
  schemas:
    Task:
      required:
      - status
      - title
      type: object
      properties:
        id:
          type: string
          example: 64ab32cd9f0c1e12a45b78d9
        title:
          type: string
          example: Preparar práctica de OpenAPI
        description:
          type: string
          example: Definir modelos y esquemas en OpenAPI
        status:
          type: string
          example: todo
          enum:
          - todo
          - in_progress
          - done
        priority:
          type: string
          example: medium
          enum:
          - low
          - medium
          - high
        dueDate:
          type: string
          format: date-time
          example: 2025-10-30T14:00:00Z
        tags:
          type: array
          items:
            type: string
            example: "[\"work\",\"urgent\"]"
    Error:
      required:
      - error
      type: object
      properties:
        error:
          type: string
          example: Validation error
        details:
          type: string
          example: The field 'title' is required
    tasks_body:
      required:
      - status
      - title
      type: object
      properties:
        title:
          type: string
          example: Preparar curso de OpenAPI
        status:
          type: string
          example: todo
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          example: "https://api.ejemplo.com/errors/validation"
        title:
          type: string
          example: "Validation error"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "Missing field 'title'"
        instance:
          type: string
          example: "/tasks"
  responses:
    ErrorResponse:
      description: Error genérico
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'